AB#565

+++ DataService/Dto/FNA/TaxLumpSumDto.cs
@@ -10,10 +10,10 @@ namespace DataService.Dto
     {
         public int Id { get; set; }
         public int FnaId { get; set; }
-        public float PreviouslyDisallowed { get; set; }
-        public float RetirementReceived { get; set; }
-        public float WithdrawalReceived { get; set; }
-        public float SeverenceReceived { get; set; }
-        public float TaxPayable { get; set; }
+        public double PreviouslyDisallowed { get; set; }
+        public double RetirementReceived { get; set; }
+        public double WithdrawalReceived { get; set; }
+        public double SeverenceReceived { get; set; }
+        public double TaxPayable { get; set; }
     }
 }
